requirements.txt
feedparser
python-telegram-bot
google-generativeai
requests
beautifulsoup4
